all field dosen dan karyawan

+++ app/Http/Controllers/Master/Karyawan/DosenController.php
@@ -10,6 +10,7 @@ use App\Models\Jurusan;
 use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
 use Illuminate\Support\Facades\Hash;
+use phpDocumentor\Reflection\Types\ArrayKey;
 
 class DosenController extends Controller
 {
@@ -55,9 +56,48 @@ class DosenController extends Controller
      */
     public function store(Request $request)
     {
-        $this->saveDosenHandler($request, 'store');
+        // $this->saveDosenHandler($request, 'store');
 
-        return redirect()->route('master.dosen.index');
+        $request['password'] = Hash::make('12345678');
+        // $dosen = Dosen::findOrFail($request);\
+        // $user = new User();
+        // dd($request->all());
+        $msg = 'Berhasil menambahkan data dosen.';
+        $status = 'OK';
+        try {
+            $user = User::create($request->all());
+            try {
+                $request['id'] = $request['id_karyawan'];
+                $user->staff()->create($request->all());
+                try {
+                    $request['id'] = $request['id_dosen'];
+                    $request['jurusan_id'] = $request['jurusan'];
+                    $user->staff->dosen()->create($request->all());
+                } catch (\Throwable $th) {
+                    $msg = 'Gagal menambahkan data dosen. Error: '.$th->getMessage();
+                    $status = 'FAIL';
+                    // $user->delete();
+                    dd($msg, $status);
+                }
+            } catch (\Throwable $th) {
+                $msg = 'Gagal menambahkan data karyawan. Error: '.$th->getMessage();
+                $status = 'FAIL';
+                $user->delete();
+                dd($msg, $status);
+            }
+        } catch (\Throwable $th) {
+            $msg = 'Gagal menambahkan data user. Error: '.$th->getMessage();
+            $status = 'FAIL';
+            dd($msg, $status);
+        }
+        // dd($request->all());
+
+        // return redirect()->route('master.dosen.index');
+        $header = ['status' => $status, 'msg' => $msg];
+        // if ($status != 'FAIL'){
+        //     return redirect('master/karyawan')->with($header);
+        // }
+        return redirect()->route('master.dosen.edit',$request->id_dosen)->with($header);
     }
 
     /**
@@ -69,6 +109,9 @@ class DosenController extends Controller
     public function edit($id)
     {
         $dosen = Dosen::where('id', '=', $id)->with(['staff.user', 'jurusan'])->get()->first();
+        if (!$dosen){
+            abort(404);
+        }
         $jurusans = Jurusan::all();
 
         return Inertia::render('Master/Karyawan/Dosen/KaryawanDosenDetail', [
@@ -86,9 +129,53 @@ class DosenController extends Controller
      */
     public function update(Request $request, $id)
     {
-        //
+        // $this->saveDosenHandler($request, 'update', $id);
 
-        $this->saveDosenHandler($request, 'update', $id);
+
+        $dosen = Dosen::findOrFail($id);
+        $user = User::findOrFail($dosen->staff->user->id);
+        // dd($user->update());
+
+        $msg = 'Berhasil menambahkan data';
+        $status = 'OK';
+        
+        try {
+            // $key = array_keys($user->toArray());
+            $user_old = $user->getOriginal();
+            $user->update($request->all());
+
+            try {
+                $staff_old = $user->staff->getOriginal();
+                $request['id'] = $request['id_karyawan'];
+                $key = array_keys($user->staff->toArray());
+                $user->staff->update($request->only($key));
+                try {
+                    // dd($request['id'], $staff_old['id']);
+                    if ($request['id'] != $staff_old['id']){
+                        // dd($request['user_id']);
+                        $user = User::findOrFail($user->id);
+                        // dd('disini');
+                    }
+                    $request['id'] = $request['id_dosen'];
+                    $request['jurusan_id'] = $request['jurusan'];
+                    $dosen_old = $user->staff->dosen->getOriginal();
+                    $key = array_keys($dosen_old);
+                    $user->staff->dosen->update($request->only($key));
+                } catch (\Throwable $th) {
+                    $msg = 'Gagal mengubah data dosen. Error: '.$th->getMessage();
+                    $status = 'OK';
+                    dd($msg, $status);
+                }
+            } catch (\Throwable $th) {
+                $msg = 'Gagal mengubah data karyawan. Error: '.$th->getMessage();
+                $status = 'OK';
+                dd($msg, $status);
+            }
+        } catch (\Throwable $th) {
+            $msg = 'Gagal mengubah data user. Error: '.$th->getMessage();
+            $status = 'OK';
+            dd($msg, $status);
+        }
 
         return redirect()->route('master.dosen.index');
     }
@@ -104,11 +191,15 @@ class DosenController extends Controller
         //
         $dosen = Dosen::findOrFail($id);
 
+        if(!$dosen){
+            abort(404);
+        }
+
         $dosen->delete();
 
-        $dosen->staff->delete();
+        // $dosen->staff->delete();
 
-        $dosen->staff->user->delete();
+        // $dosen->staff->user->delete();
 
         return redirect()->route('master.dosen.index');
     }
